Split CoreExtensions into separate files per functionality
